Moved ring searching to view package
