+++ core/view/search.go
@@ -1,4 +1,4 @@
-package core
+package view
 
 import (
 	"errors"
